% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_gdata.R
\name{build_gdata}
\alias{build_gdata}
\title{Generate global parameters dataset}
\usage{
build_gdata(
  n_ctcH_PSA = 2, t_ctcH_PSA = 10 / 60 / 24, n_ctcP_PSA = 0, t_ctcP_PSA = 5 / 60 / 24,
  n_ctcH_PW = 4, t_ctcH_PW = 15 / 60 / 24, n_ctcP_PW = 4, t_ctcP_PW = 30 / 60 / 24,
  n_ctcH_H = 5, t_ctcH_H = 3 / 60 / 24, t_ctcV_PW = 20 / 60 / 24,
  I = 185 / 100000, d = 10, R0 = 1.29, tw = 35,
  tSA  = 2 / 24, tISO = 11, tIC  = 15, tSL  = 14, tESL = 28, tE  = 5,
  tEA = 2, tES = 2, tIA = 7, tIM = 8, tIS = 9, tLI = 60, tHI = 150,
  epsPPSA = 0.1, epsHPSA = 0.1, epsHPW = 0.1, epsPPW = 0.1,
  epsVPW = 0.1, epsPHSA = 0.1, epsPHW = 0.1, epsHHW = 0.1,
  ttestSA = 2 / 24, ttestPW = 2 / 24, ttestHW = 2 / 24, ttestsymp = 2 / 24,
  tbtwtestP = 14, tbtwtestH = 30, tbeftestPsymp = 2 / 24, tbeftestHsymp = 1,
  psympNI = 0.5, psympLI = 0.2, psympHI = 0.1, psevNI = 0.5, psevLI = 0.3, psevHI = 0.1,
  pISO = 1, pSL = 0.3, pESL = 1, pSLT = 0.01, pIC = 0.3, pdieIC = 0.005,
  pLI = 0.20, pHI = 0.5, hNI2LI = 1 / 30, hLI2HI = 1 / 60, rinfLI = 0.7, rinfHI = 0.5,
  rsymp = 1, rsev = 1,
  ptestPSAsymp = 1, ptestPSANI = .75, ptestPSALI = 0.50, ptestPSAHI = 0.1,
  ptestPWsymp = 0.95, ptestPWNI = 0.75, ptestPWLI = 0.50, ptestPWHI = 0.10,
  ptestHsymp = 0.85, ptestHNI = 0.75, ptestHLI = 0.50, ptestHHI = 0.20,
  senSA = 0.85, speSA = 0.95, senW = 0.85, speW = 0.95,
  senH = 0.85, speH = 0.95, sensymp = 0.85, spesymp = 0.95)
}
\arguments{
\item{n_ctcH_PSA}{a positive number, the daily number of contacts with professionals per patient in the screening area}

\item{t_ctcH_PSA}{a positive number, the average duration of contact between professional and patient per day in the screening area (in days, ex: if 15 min, t_ctcH_PSA = 15/60/24)}

\item{n_ctcP_PSA}{a positive number, the daily number of contacts with other patients per patient in the screening area}

\item{t_ctcP_PSA}{a positive number, the average duration of contact between patients per day in the screening area (in days)}

\item{n_ctcH_PW}{a positive number, the number of contact between professional and patient per day in the ward}

\item{t_ctcH_PW}{a positive number, the average duration of contact between patients per day in the ward (in days)}

\item{n_ctcP_PW}{a positive number, the number of contacts with other patients per patient per day, in the ward}

\item{t_ctcP_PW}{a positive number, the average duration of contact between patients per day in the ward (in days)}

\item{n_ctcH_H}{a positive number, the number of contacts with professionals per professional per day}

\item{t_ctcH_H}{a positive number, the average duration of contact between professionals in the ward (in days)}

\item{t_ctcV_PW}{a positive number, the average duration of contact between patients per day in the ward (in days)}

\item{I}{a positive number, the daily incidence}

\item{d}{a positive number, the average disease duration (in days)}

\item{R0}{a positive number, the basic reproduction number}

\item{tw}{a positive number, the average number of working hours per week (in hours)}

\item{tSA}{a positive number, the average duration (in days) before full admission (in screening area for clinical exam, administrative procedure, etc)}

\item{tISO}{a positive number, the average duration (in days) of confinement (isolated ward or contact restriction)}

\item{tIC}{a positive number, the average duration (in days) of stay in intensive care}

\item{tSL}{a positive number, the average duration (in days) of sick leave}

\item{tESL}{a positive number, the average duration (in days) of extended sick leave}

\item{tE}{a positive number, the average duration (in days) of epidemiological state E (exposed - non contagious)}

\item{tEA}{a positive number, the average duration (in days) of epidemiological state EA (exposed - contagious pre-asymptomatic)}

\item{tES}{a positive number, the average duration (in days) of epidemiological state ES (exposed - contagious pre-symptomatic)}

\item{tIA}{a positive number, the average duration (in days) of epidemiological state IA (infectious asymptomatic)}

\item{tIM}{a positive number, the average duration (in days) of epidemiological state IM (infectious with mild symptoms)}

\item{tIS}{a positive number, the average duration (in days) of epidemiological state IS (infectious with severe symptoms)}

\item{tLI}{a positive number, the average duration (in days) of low immunity persistence before return to non immune status}

\item{tHI}{a positive number, the average duration (in days) of high immunity persistence before return to low immune status}

\item{epsPPSA}{a positive ratio, the average infection-reducing ratio for patient infection during patient-to-patient contacts in the screening area (0 = no protection, 1 = complete protection/no infection possible)}

\item{epsHPSA}{a positive ratio, the average infection-reducing ratio for patient infection during professional-to-patient contacts in the screening area (0 = no protection, 1 = complete protection/no infection possible)}

\item{epsHPW}{a positive ratio, the average infection-reducing ratio for patient infection during professional-to-patient contacts in the ward (0 = no protection, 1 = complete protection/no infection possible)}

\item{epsPPW}{a positive ratio, the average infection-reducing ratio for patient infection during patient-to-patient contacts in the ward (0 = no protection, 1 = complete protection/no infection possible)}

\item{epsVPW}{a positive ratio, the average infection-reducing ratio for patient infection during visitor-to-patient contacts (0 = no protection, 1 = complete protection/no infection possible)}

\item{epsPHSA}{a positive ratio, the average infection-reducing ratio for professional infection during patient-to-professional contacts in the screening area (0 = no protection, 1 = complete protection/no infection possible)}

\item{epsPHW}{a positive ratio, the average infection-reducing ratio for professional infection during patient-to-professional contacts in the ward (0 = no protection, 1 = complete protection/no infection possible)}

\item{epsHHW}{a positive ratio, the average infection-reducing ratio for professional infection during professional-to-professional contacts (0 = no protection, 1 = complete protection/no infection possible)}

\item{ttestSA}{a positive number, the average duration (in days) of test in the screening area (from decision to test to action - such as transfer -  after obtaining the test result)}

\item{ttestPW}{a positive number, the average duration (in days) of test in the ward for patient screening (from decision to test to action - such as transfer -  after obtaining the test result)}

\item{ttestHW}{a positive number, the average duration (in days) of test for professional screening (from decision to test to action - such as sick leave -  after obtaining the test result)}

\item{ttestsymp}{a positive number, the average duration (in days) of test for symptomatic individuals (from decision to test to action - such as sick leave or transfer -  after obtaining the test result)}

\item{tbtwtestP}{a positive number, the average duration (in days) between two patient screening testing-events in the ward}

\item{tbtwtestH}{a positive number, the average duration (in days) between two professional screening testing-events in the ward}

\item{tbeftestPsymp}{a positive number, the average duration (in days) of symptomatic patients detection (before test)}

\item{tbeftestHsymp}{a positive number, the average duration (in days) of symptomatic professional detection (before test)}

\item{psympNI}{a positive probability, the conditional probability to be symptomatic when non immune}

\item{psympLI}{a positive probability, the conditional probability to be symptomatic despite a low immunity}

\item{psympHI}{a positive probability, the conditional probability to be symptomatic despite a high immunity}

\item{psevNI}{a positive probability, the conditional probability to develop severe symptoms when symptomatic and non immune}

\item{psevLI}{a positive probability, the conditional probability to develop severe symptoms when symptomatic and despite a low immunity}

\item{psevHI}{a positive probability, the conditional probability to develop severe symptoms when symptomatic and despite a high immunity}

\item{pISO}{a positive probability, the probability to be isolated if confirmed case}

\item{pSL}{a positive probability, the probability for professional with mild symptoms of taking sick leave}

\item{pESL}{a positive probability, the probability for professional with severe symptoms  of taking extended sick leave}

\item{pSLT}{positive probability, the additional probability for professionals of taking sick leave after positive test}

\item{pIC}{a positive probability, the probability for patient with severe symptoms to be transfer in intensive care}

\item{pdieIC}{a positive probability, the probability to die in intensive care}

\item{pLI}{a positive probability, the probability to have low immunity at the admission (proportion of individuals with low immunity in the community)}

\item{pHI}{a positive probability, the probability to have high immunity at the admission (proportion of individuals with high immunity in the community)}

\item{hNI2LI}{a positive probability, the daily probability to gain low immunity when non immune}

\item{hLI2HI}{a positive probability, the daily probability to gain high immunity when having low immunity}

\item{rinfLI}{a positive ratio, the average infection-reducing ratio for individuals with low immunity compared to non immune (can be interpreted as low immunity efficiency)}

\item{rinfHI}{a positive ratio, the average infection-reducing ratio for individuals with high immunity compared to non immune (can be interpreted as high immunity efficiency)}

\item{rsymp}{a ratio, the average ratio increasing or reducing the probability to develop symptoms for patients compared to general population (professionals)}

\item{rsev}{a ratio, the average ratio increasing or reducing the probability to develop severe symptoms for symptomatic patients compared to general population (professionals)}

\item{ptestPSAsymp}{a positive probability, the probability to test symptomatic patients in the screening area}

\item{ptestPSANI}{a positive probability, the probability to test non immune patients in the screening area}

\item{ptestPSALI}{a positive probability, the probability to test patients with low immunity in the screening area}

\item{ptestPSAHI}{a positive probability, the probability to test patients with high immunity in the screening area}

\item{ptestPWsymp}{a positive probability, the probability to test symptomatic patients in the ward (Diagnostic testing)}

\item{ptestPWNI}{a positive proportion, the proportion of non immune patients tested during screening testing in the ward}

\item{ptestPWLI}{a positive proportion, the proportion of patients with low immunity tested during screening testing in the ward}

\item{ptestPWHI}{a positive proportion, the proportion of patients with high immunity tested during screening testing in the ward}

\item{ptestHsymp}{a positive probability, the probability to test symptomatic professionals (Diagnostic testing)}

\item{ptestHNI}{a positive proportion, the proportion of non immune professionals tested during screening testing}

\item{ptestHLI}{a positive proportion, the proportion of professionals with low immunity tested during screening testing}

\item{ptestHHI}{a positive proportion, the proportion of professionals with high immunity tested during screening testing}

\item{senSA}{a positive probability, the sensibility of tests use within the screening area}

\item{speSA}{a positive probability, the specificity of tests use within the screening area}

\item{senW}{a positive probability, the sensibility of tests use for patient screening in the ward}

\item{speW}{a positive probability, the specificity of tests use for patient screening in the ward}

\item{senH}{a positive probability, the sensibility of tests use for professional screening}

\item{speH}{a positive probability, the specificity of tests use for professional screening}

\item{sensymp}{a positive probability, the sensibility of tests use for symptomatic individuals}

\item{spesymp}{a positive probability, the specificity of tests use for symptomatic individuals}
}
\value{
Named list of parameters
}
\description{
It calculates 1) the prevalence from daily incidence in the community and average disease duration,
2) the probability of transmission per day spent in contact based on R0 and average disease duration,
3) the time ratios of the epidemiological state based on average states duration,
4) the proportion of time outside of work, and
5) the proportion of probabilities of developing mild and severe symptoms for patients.

Then the functions formats all global parameters required for the transitions as a named list readable by SimInf package.
}
\examples{
build_gdata()

}
