% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/keyoutput.R
\name{keyoutput}
\alias{keyoutput}
\title{Generate summaries of trajectories}
\usage{
keyoutput(trajmwss,  scale = 0, focus = NULL, outb_Thhold = 1)
}
\arguments{
\item{trajmwss}{mwss object. List of list produced by \code{multisim} function}

\item{scale}{NULL or numeric vector. 0 is the facility scale, 1 is the ward scale.}

\item{focus}{String. Limit the output to infection data (\code{focus = "infections"}), incidence data (\code{focus = "incidence"}) or test counter  (\code{focus = "test"}).}

\item{outb_Thhold}{Integer. Defines the minimal number of cases requested to define an epidemic (default is 1, only used with \code{scale = 1}).}
}
\value{
List
}
\description{
\code{keyoutput} apply to a mwss object returns a list with number of simulation (\code{nSimulations}), number of wards (\code{nwards}),
the number of admission over the whole simulation (\code{nadm}),
the maximal number of patients that were simultaneously isolated over the whole simulation (\code{ISO}),
the maximal number of professionals that simultaneously took sick leave over the whole simulation (\code{SL}),
the statistical summary of infections over the whole simulation (\code{infection}),
the statistical summary of incidence over the whole simulation (\code{incidence}),
the statistical summary of tests implemented over the whole simulation (\code{tests}).

Used with the \code{scale = 0}, the function returns statistics calculated at the facility scale.
Used with the \code{scale = 1}, the function returns statistics calculated for each ward.

Using focus argument limit outputs to one of the three last list contains in the generic method.

\code{n_extinction} is the number of extinction among all simulations. \code{perc_outbreak} is the proportion of simulations were an epidemic was observed (only with \code{scale = 1}).
\code{n_noso} is the average number of nosocomial infections in a given population over the simulations.
\code{n_intro} is the average number of infected patients introduced in a given population over the simulations.
\code{n_out} is the average number of professionals of a given population who was infected in community over the simulations.
\code{incPA, incPM, incPS, incHA, incHM} and \code{incHS} are respectively the average incidence of asymptomatic, mild symptomatic and severe symptomatic patients and professionals (H: healthcare workers) over the simulations.
\code{incP} and \code{incH} are the respective incidence among patients and professionals (including asymptomatics and symptomatics).
\code{nTestP} and \code{nTestH} are the respective number of tests performed on patients and professionals.
\code{n_dtestP} and \code{n_dtestH} are the respective average daily number of tests performed on patients and professionals.
}
\examples{
data("toydata")
list2env(toydata,envir=.GlobalEnv)
gdata <- build_gdata()
model <- mwss(ward_names, pop_size_P, pop_size_H, nVisits, LS, gdata, tSim = 30)
results <- multisim(model, 5, ward_names)

keyoutput(results,  scale = 0, focus = NULL)
keyoutput(results,  scale = 1, focus = NULL, outb_Thhold = 1)
keyoutput(results,  scale = 0, focus = "incidence")

}
