% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multisim.R
\name{multisim}
\alias{multisim}
\title{Simulate multiple trajectories}
\usage{
multisim(model, nSim, ward_names)
}
\arguments{
\item{model}{SimInf model generated by the function \code{mwss}}

\item{nSim}{Number of simulations}

\item{ward_names}{String vector with ward names/id in the same order than the one used in the function \code{mwss} to build the model}
}
\value{
List of data.table of class 'mwss'
}
\description{
The function \code{multisim} run the model multiple times, save the results in a list of class 'mwss'.
Results are trajectories simulated by SimInf package. Epidemiological and immunity states of each ward are save at each time steps.
When run in shiny, a progress bar is also displayed during simulations.
}
\examples{
data("toydata")
list2env(toydata,envir=.GlobalEnv)
gdata <- build_gdata()
model <- mwss(ward_names, pop_size_P, pop_size_H, nVisits, LS, gdata, tSim = 30)
results <- multisim(model, 5, ward_names)

}
