% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_daily_cases.R
\name{plot_daily_cases}
\alias{plot_daily_cases}
\title{Plot the incidence}
\usage{
plot_daily_cases(trajmwss)
}
\arguments{
\item{trajmwss}{List of data.table. Epidemiological trajectories simulated by the function \code{mwss::multisim}}
}
\value{
Incidence plot.
}
\description{
The function \code{plot_daily_cases} returns FIX ME
}
\examples{

data("toydata")
list2env(toydata,envir=.GlobalEnv)
gdata <- build_gdata()

model <- mwss(ward_names, pop_size_P, pop_size_H, nVisits, LS, gdata, tSim = 30)
results <- multisim(model, 5, ward_names)

plot_daily_cases(trajmwss = results)

}
