% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_incidence.R
\name{plot_incidence}
\alias{plot_incidence}
\title{Plot the cumulative incidence}
\usage{
plot_incidence(
          trajmwss, scale = 0,
          pop = FALSE, iter = FALSE,
          ward = FALSE, display_sd = TRUE)
}
\arguments{
\item{trajmwss}{List of data.table. Epidemiological trajectories simulated by the function \code{mwss::multisim}}

\item{scale}{Integer. Indicated the scale for statistics: 0/ whole facility, 1/ wards.}

\item{pop}{String. Define a subpopulation for statistics ("P":Cumulative incidence among patients, "Psymp": Cumulative incidence among patients (only considering symptomatics), "H": Cumulative incidence among professionals, "Hsymp": Cumulative incidence among professionals (only considering symptomatics); default is FALSE: considering all individuals).}

\item{iter}{Integer. Define a specific simulation. Default is FALSE, providing average values over all simulations.}

\item{ward}{String. Define a specific ward. Default is FALSE, considering all wards. At the facility scale '0' no ward can be specified.}

\item{display_sd}{Logical. Determine standard deviation display. Default is TRUE.}
}
\value{
Cumulative incidence plot.
}
\description{
The function \code{plot_incidence} returns a barplot representing the average cumulative incidence over simulations at different scales and for different populations
}
\examples{

data("toydata")
list2env(toydata,envir=.GlobalEnv)
gdata <- build_gdata()

model <- mwss(ward_names, pop_size_P, pop_size_H, nVisits, LS, gdata, tSim = 30)
results <- multisim(model, 5, ward_names)

plot_incidence(trajmwss = results)

}
