% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/randomContacts.R
\name{randomContacts}
\alias{randomContacts}
\title{Generate random matrix of contact}
\usage{
randomContacts(pop_size_H, ward_names)
}
\arguments{
\item{pop_size_H}{Numerical vector. Contains the number of professional affiliated to each ward in the same order than ward_names argument.}

\item{ward_names}{String vector. Contains the ward names in the same order than pop_size_H argument.}
}
\value{
List of two data.frame.
sharedTime: proportion of time spent in each ward (row) for each professional (column) ;
contactMat: average time spent in each ward (column) by professionals of each ward (row).
}
\description{
\code{randomContacts} returns a list with random contact between wards.
}
\examples{
data("toydata")
list2env(toydata,envir=.GlobalEnv)

randomContacts(pop_size_H, ward_names)

}
