% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toydata.R
\docType{data}
\name{toydata}
\alias{toydata}
\title{Toy data for mwss examples}
\format{
An object of class \code{"list"}.
}
\usage{
data(toydata)
}
\description{
Toy data gerenated for mwss examples contains baseline data for a 5-ward facility.
}
\examples{
data(toydata)
str(toydata)
nwards <- toydata$nwards
ward_names <- toydata$ward_names
\donttest{
list2env(toydata,envir=.GlobalEnv)
matContact <- randomContacts(pop_size_H, ward_names)$contactMat
plot_connectivity(matContact, pop_size_P, verbose = FALSE)
}
}
\keyword{datasets}
