% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_network.R
\name{build_network}
\alias{build_network}
\title{Creates a synthetic hospital ward network}
\usage{
build_network()
}
\arguments{
\item{n_buildings}{integer. number of buildings (="clusters")}

\item{n_wards}{numerical vector. Vector of length (n_buildings) with number of wards for each building}

\item{total_patients}{integer. Total number of patients in the hospital}

\item{total_HCW}{integer. Total number of HCW in the hospital}

\item{minLS}{integer. Minimal number of days for random length of stay}

\item{maxLS}{integer. Maximal number of days for random length of stay}

\item{within_clust}{numeric. Ratio of intra-building clustering (relative to inter-building clustering)}

\item{silent}{logical. Print plot if TRUE}
}
\value{
A list of 9 elements: 1. ward_names, 2. pop_size_P, 3. pop_size_H, nVisits, 5. LS (length of stay), 6. Hplanning, 7. matContact, 8. IMMstate, 9. EPIstate
}
\description{
This function creates a synthetic hospital ward network with additional information on the population size, visitors etc.
You can use this script to generate different input data to be used as examples of "network structures".
}
